cleanup tests, add changelog entry
