+++ CHANGELOG.md
@@ -6,12 +6,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-* add `gpx` file to demo activities
+* Add `gpx` file to demo activities.
+* Use [dask](https://docs.dask.org/en/) to parse activity files in parallel to speed
+  up file import / reimport.
 ### Changed
 * Show "made with ❤️" in footer only on settings and help page.
 * Moved `tests` folder from `workoutizer/wkz/tests/` to `workoutizer/tests` root folder.
 * Run all tests together in github actions `Test` and `Release` pipeline.
-* ensure activity leaflet map has always height of > 400 pixels 
+* Ensure activity leaflet map has always height of > 400 pixels
 
 ## [0.18.0](https://github.com/fgebhart/workoutizer/releases/tag/0.18.0) - 2021-05-16
 ### Added
